// packages/codegen-kernel/src/pluginTypes.ts
export type Framework = "Flutter" | "SwiftUI" | "HTML" | "Tailwind" | "Compose";
// ...
export type SolidColorConversion = {
  hex: string;
  colorName: string;
  exportValue: string;
  contrastWhite: number;
  contrastBlack: number;
  meta?: string;
};
export type LinearGradientConversion = {
  cssPreview: string;
  exportValue: string;
};